Add files via upload
Three part project intended to introduce us to dealing with data that is stored in CSV format using pandas. The data is contained in one file, covering the time period from the late 1978, when satellite coverage first went online. 

part 1 Loaded csv data into pandas, cleaned it, reformatted it and wrote two new files to disk covering 1979-2017 and 2018.

part 2 In this project, I reproduced figures from the web created by the National Snow and Ice Data Center (NSIDC) and Japan's version, the Arctic Archive System (ADS). 

Part 3 The last part of this project was intended to analyze the Arctic sea ice extent data and use analysis to predict the date the ice will disappear altogether. I did this by fitting lines to the March and September data. This is called simple linear regression, calculated using ordinary least squares. Recalculated the data as an anomaly. The anomaly is the difference between a given data point and the mean. I was interested in the maximum ice extent, which occurs in March but varies from year to year as to the exact day, and the minimum ice extent, which happens in September. Therefore, I was looking at those months as single entities, example the March and September means for each year. In other words, instead of looking at daily data, I looked at monthly data, I calculated the difference between the monthly mean for a given year and the mean of the means for all of the years. For example, I calculated the difference the between the mean of the March data for 1981 and the mean of the March means for all the years looking at the anomaly instead of the absolute values that allowed me to create a better visualization that compares how ice behavior is changing in March with how ice behavior is changing in September.